Committing handles for vectors, linear operators, and vector spaces.
Notes:
(1) There is a pending bug with indexing into block vectors. Ross is working on this.

(2) I have knowingly used a non-standard version of automake here; I'm
working off of a loaner laptop that doesn't have the right versions of automake.
Someone please re-bootstrap with the approved version.

(3) test_block_op and test_handles need to be added to the test harness defintiions.