changed some comments in ModalSolverUtils
SVQBOrthoManager was failing to detect rank degeneracy in test multivectors and
causing problems in projectAndNormalize() tests; i modified the heuristics but
will need to talk with rich/ulrich about what should be done here.

modified both orthomanager tests to take --verbose and --debug, and modified the test harness
definition file to this effect.