----------------------------------------------------------------------  Modified Files:  	paper.bib paper.tex referees.txt  ----------------------------------------------------------------------
Completed my changes to this paper and updated the comments to the referees.

Ken: At this point there are just a few questions for you to address that should be
fairly straight-forward.  Please try to complete at your earliest convenience so we
can get this paper out the door.

Thanks.

Mike