(1) Support for complex-valued expressions
(2) Additional transformations to simplify expressions:
(a) removing multiplication by 1
(b) processing unary minus operators, for example, -(-x) --> x and
x + (-y) --> x-y
(c) ordering terms in a sum to identify cancellations. This is important
in complex-valued calculations because it allows exact cancellations of
the imaginary part of expressions that should reduce to reals, such as
z*conj(z).