Changes for Bug 2546
1. Major changes to nox's linear system AztecOO object to support reuse of the preconditioner.

Eliminated the linear solver parameter "Reuse Preconditioner" in the AztecOO linear system object and replaced it with the following option:

"Preconditioner Reuse Policy" = "Reuse"
"Preconditioner Reuse Policy" = "Rebuild"
"Preconditioner Reuse Policy" = "Recompute"

The first option, "Reuse" just blindly reapplies the prconditioner from the last evaluation.  The "Rebuild" option always completely destroys and then rebuilds the preconditioner each time a linear solve is requested.  The third option, "Recompute" will recompute the preconditioner, but will try to efficiently reuse any objects that don't need to be destroyed.  How efficient the "Recompute" option is depends on the type of preconditioner.  For example if we are using ILU in Ifpack, the graph that is allocated does not need to be destroyed and recomputed even though the computation will recompute the factorization when the matrix changes.  These changes support native Aztec, Ifpack and ML preconditioners.

This required alot of changes as to how nox handles the AztecOO object and when it sets the operators.  The order of operations is important in most cases.  This also uncovered a bug in aztecoo that will hopefully be fixed for release 7.0.  I currently have a workaround in place.

This also affected the loca groups and especially, the householder algorithms that replace the jacobian and preconditioner operators in the linearsystem_aztecoo object.  I fixed all of this for the nox/loca tests, but if there are more cases not in the test section, please check these asap.

If there are fixes to be made to the LinearSyztem_AztecOO object - Please - talk to me before you make any changes.  I found a few changes made in the last year that broke some of the assumptions for the Group/Linear System that cause problems for us in Charon.

I have not completed all the changes to the documentation. Those will be added tomorrow.

2. Added 8 new tests that should catch all permutations of the preconditioning options and reveal any changes to aztec or ifpack that would impact our epetra support.

3. Updated some documentation.

4. Added a few headers to the new epetra DebugUtils that were missing in parallel, and removed a std library include from the header file.  NOTE: All std library includes should be brought in thru the nox config file so it is portable!

All tests pass in serial and parallel...