Bug 2841
I wrote some comments at the beginning that explain what this class will do.
It will be almost identical to the linear interpolation except that the actual
interpolating polynomial will include xdot information in addition to x
information and will be 2nd order in x and 1st order in xdot.