Here is what I have done as part of this checkin:
* I have added another constructor to
Teuchos::StringToIntegralParameterEntryValidator that allows the client to set
an array of documentation strings for each valid string value.  This allow for
better strucuturing of parameter documentation and more automatic
maintainance.  This is something that Ken Stanley had asked for and I have
found it to be needed.

* I have updated the valid parameter lists for
Thyra::DirectionalFiniteDiffCalculator and MoochoPack::MoochoThyraSolver to
use the new string value documentation feature in the
Teuchos::StringToIntegralParameterEntryValidator class.

* I added the machinary to automatically print out the valid parameter list
for MoochoPack::MoochoThyraSolver and have it automatically incorporated into
the doxygen documentation.