Adding examples for LOBPCG and block Davidson to read in HB or Matrix Market files describing an eigenproblem (standard or generalized) and solve the eigenproblem.  This is a good example for users and utility for developers. The examples require a utility from EpetraExt, so there is now a configure-time check to see if EpetraExt is enabled.  If so, these examples will be compiled.
The package was bootstrapped and the introduction of the configure check introduced
a new variable to each Makefile.in, that is why they all need to be checked in.

Block Krylov-Schur already had an example that would read in a matrix from a HB
file and solve a standard eigenproblem.  It has been updated to use EpetraExt like
the other new examples, thus the file was renamed to something more generic.