This commit is the first shot at resurrecting the NOX wrappers in the development branch:
* A refactor has been applied that better handles the nested
  namespace structure of the NOX package.

* Support for NOX::ParameterList has been eliminated.

* Support for Teuchos::ParameterList has been added.

* Support for Teuchos::RefCountPtr has been added.  If the C++
  code expects an object wrapped inside a Teuchos::RefCountPtr,
  the corresponding python code will expect just a simple object
  and will apply the RefCountPtr internally.  This is implemented
  with macros and typemaps in the Teuchos interface file, but the
  macros are applied as needed in the NOX interface files.

There are certain occasions when a ParameterList created in
python will raise an exception when one of its parameters is
attempted to be accessed from C++.  It appears that the
Teuchos::any_cast *should* be successful, but it is not.  I'm
looking into it.  This problem makes it currently impossible to
get a Techos::RefCountPtr<Teuchos::ParameterList> to work as a
function/method/constructor argument.