Several changes:
1)  Added max/min operators to Fad and Taylor classes.  This requires including
<algorithm>.
2)  Added configure checks for additional headers like new, map, vector,
valarray, iterator, typeinfo, etc... that Sacado is now using and added the
includes to Sacado_ConfigDefs.h.  Removed the raw includes from many of the
Sacado source files.
3)  Changed the GeneralFad classes to put the value in the storage class
instead of directly in the GeneralFad class.  This is to pave the way for
a VectorFad class where the value will be stored elsewhere.