Removed the inclusion of "Teuchos_config.h" in the header.  This header was not necessary and only causing compiler warning from the redefinitions of "PACKAGE", "VERSION", etc.
Fixed the logic in the TEST_FOR_EXCEPTION macros in both the start and stop
calls for the timer.  Actually stopped the timer in Teuchos_stopTimer :)

This has been integrated into the GMRES solver in AztecOO and seems to work well.
I'll update the Time test so that this tool gets exercised in the nightly tests.