This is a pretty big checkin.
Here I have added a example for how to use Thyra/Stratimikos to quickly build
specialized preconditioners.  This example lives in:

  packages/stratmikos/example

and is called MixedOrderPhysicsBasedPreconditioner.cpp.  See
stratimikos/test/definition for how to run the example with test matrices and
various solver building blocks.

Associated with this example, I made many improvements to the introspection of
Thyra-based objects.  I also make numerous other improvements.

Here is what I did more specifically:

1) Added a new base class Teuchos::LabeledObject that allows a client to
query and set an object's string label.

2) Derived Teuchos::Describable from Teuchos::LabeledObject and updated the
default implementations of description() and describe() to show the object's
label if set.

3) Updated Thyra::AztecOOLinearOpWithSolveFactory to allow the client-set
preconditioner operator to not directly support the Thyra::EpetraLinearOpBase
interface.  I used the function Thyra::makeEpetrWrapper(...) written by Kevin
Long for AztecOO to use with Meros to make this possible.  The AztecOO/Thyra
wrappers are messy and this is not pretty (please bring use Belos :-)).

4) Added non-member constructor functions for Thyra::EpetraLinearOp called
Thyra::epetraLinearOp(...).

5) Incorporated the updated EpetraExt_CrsMatrixIn.cpp code that Mike Heroux
sent me.  This allows you to read in rectangular matrices in Matrix Market
format.

6) Improved the implementations of the overrides of
Teuchos::Describable::describe(...) and the verbose output (due to the
Teuchos::VerboseObject base class) for may of the classes.  More work is needed
to make the output more consistent.

7) I added helper functions in Teuchos_VerboseObjectParameterListHepers.hpp to
make it very easy for a subclass of Teuchos::VerboseObject and
Teuchos::ParameterListAcceptor to set it's verbosity settings from a parameter
sublist.  Basically, only two lines of code need to be added to a current such
subclass to enable this feature.  I also updated the
Teuchos::VerboseObject[Base] class to allow the verbosity level and the output
stream that is set by the parameter list to not be overridden by current code.
This allows the user to set a verbosity level and/or redirect output to a file
in a way that overrides anything set in code.  The file-based output does not
handle output in parallel yet so there is still more work to be done and it is
not 100 percent clear how to do this in general.

8) I have started to reformat some of the files according to the new Thyra
formatting guidelines.

I am sure that I have made other small changes as well ...