This check-in represents the refactoring of scaling support code for EpetraExt/Thyra ModelEvaluators for the ASC Vertical Integration Milestone.
Here is what I did:

1) I created new files EpetraExt_ModelEvaluatorScalingTools.[h,cpp] that
contains the beginnings of tools to handle scaling the objects in the
EpetraExt::ModelEvaluator::[In,Out]Args containers.  This will provide a
common, extensible set of tools scaling that can be used by Thyra, LOCA, and
any other nonlinear algorithms.  Right now these tools only support scaling
w.r.t. x and f but in the very near future we will add support for scaling of
p(l) and g(j).  These is now great infrastructure for being able to expand the
scaling support very quickly.

2) I refactored Thyra::EpetraModelEvaluator to use the new
EpetraExt::ModelEvaluator scaling tools.  This has cleaned up this class a lot
but there is still lots of work to do to clean this class up more.  For
example, there is still code that explicitly computes inverse row-sum scalings
for the Jacobian to use to scale f.  All of the scaling support needs to be
factored out into separate objects to make Thyra::EpetraModelEvaluator smaller
and more flexible and to allow for more reuse outside of Epetra.  I also
removed the 'Do State Variable Scaling' parameter since this did not work well
at all.

3) I fixed a bug in Teuchos::Parameter[List,Entry] where setting a sublist was
not being recognized as a sublist.  This occurred in the new
setParametersNotAlreadySet(...) function and was discovered in Moocho Charon.
To fix the problem, I removed the data member isList_ from
Teuchos::ParameterEntry.  Now, the isList() function just checks the type of
the entry and if it is a PL then it is a PL.  If it talks like a duck and
walks like I duck it is a duck.  Just because it wares a sign that says duck
does not make it a duck!  This data member was completely unnecessary and
complicated maintenance.

4) I added the beginnings of a general interface for setting up floating point
traps.  This code is based on code that Eric Phipps gave me that come from f2c
(which was originally written by David Gay).  I added this to help in tracking
down the cause of NaNs etc. and I tried using it in Charon but it did not
catch the problem (which was actually caused by a memory usage error, not a
floating point error).  I have not added a test for this code in Teuchos yet
but I am checking it in just to get an idea of how portable it is.  If there
is a build failure on some platform, I will remove this from the makefile and
get back to this later.  I think this is potentially useful tool that should
be a welcome addition to Teuchos and should make debugging NaNs easier down
the line.



Charon checkin message:
-----------------------

This is a small but very important checkin in support of the ASC Vertical
Integration Milestone.  I turns out that the refactoring that created
MoochoCharonSolver on 2007/04/09 resulted in any exception being thrown from
MOOCHO result in a pass through to NOX.  Some of my tests where badly written
where they actually passed when I did this.  This resulted in a latter checkin
to break some tests but I never know it.  It turned out that the actual
problem was due to trying to set a Charon parameter vector when there where no
parameters.  This resulted in uninitialized memory being read.  One of the
test cases just resulted in NaNs being produced but valgrind reported no
errors.  Thank goodness that valgrind did complain on one of the other tests
which allowed me to find the cause of the problem.

One note, the problem was that I did a paramNames_[0] where
paramNames_.size()==0 and paramNames_ was an std::vector<std::string> object.
I used std::vector<...> since this is what Charon code accepted.  If I had
used Teuchos::Array<std::vector> instead, I would have gotten a nice runtime
error thrown right away and I would have saved at least half a days work.

This is just yet another example of why we need to get away from using raw
pointers and direct access to STL containers and that valgrind can be
worthless in many cases in finding memory usage errors.  What we need are the
Tee Thyra Coding
Guidelines document.

The other bug I fixed was that I had to put an if statement around the output
of the final observation values (the currents in g(0)) when there was no g(0)
supported.

All of the tests now pass.