Adding a MultiPoint capability which creates a "Super" EpetraExt::ModelEvaluator from another EpetraExt::ModelEvaluator, where a set of steady problems are solved simultaneously. With this extra layer, optimization algorithms for steady problems can now do multi-point optimization problems. Parallelism over the multiple points, and formation of the larger block system, uses the "4D" infrastructure in the epetraext/src/block directory.
Also being checked in is a model evaluator for a simple test
problem, which is just a small extension of the 4Dopt problem
(in epetraext/examples/model_evaluator/4dopt) and the
driver for the multipoint optimization problem
(in moocho/thyra/example/NLPThyraEpetraModelEval4DOpt).
The example solves a 20-point multi-point problem and gets
the same answer on 1-20 processors.