This is a quite big check-in.  Here I am essentially committing, in one big bang, all of the software needed to implement a Thyra-based multi-period ModelEvaluator class.
Here are some of the more significant items:

1) The example program MultiPeriodNLPThyraEpetraAdvDiffReactOptMain.cpp shows
the use of the new MultiPeriod ModelEvaluator software and it is based on the
GL problem of Denis Ridzal.

2) The class Thyra::DefaultMultiPeriodModelEvaluator is finally finished for
use with a single cluster of processes.  This class uses an implicit block
diagonal LOWSB object to implement W.  This class is a good example of the
power of Thyra and this is a true ANA class.  Extending this class for
multiple clusters of processes should not be too difficult to do but it will
degrade the ANA nature of the software somewhat.  This class depends on a lot
of new Thyra operator/vector software described below.

3) The extended ANA operator/vector interfaces
Thyra::BlockedLinearOpWithSolveBase and
Thyra::PhysicallyBlockedLinearOpWithSolveBase have been added.  These derive
from the current blocked LOB interfaces and allow a client to access and
construct a LOWSB object as smaller sub-blocks.  It is yet to be seen how
general or efficient these interfaces will be long term but this is a good
start.

4) I have began the implementation of a
Thyra::DefaultBlockedTriangularLinearOpWithSolve class that will eventually
handle all implicit upper or lower blocked triangular LOWSB objects where the
block diagonals are LOWSB objects.  The current implementation only handles
the LOWSB diagonal blocks and not the off-diagonal LOB objects.  Also, this
class will only work safely for a single cluster of processes and does not
accept any user-defined solve criteria yet.  Code and some type of Comm needs
to be added to handle the multiple-cluster case in that if any block diagonal
solve fails on any cluster then failure needs to be reported on all of the
clusters.

5) I have finally created a real Thyra::DefaultProductMultiVector concrete
subclass that does everything that it should do.  There is still the issue
that this will not be quite as efficient as a purpose-built SPMD product
MultiVector class since it will perform numBlocks calls to MPI_Allreduce(...)
instead of just one.

The above are the major things that have been changed/added.  There are also a
lot of other smaller changes/updates:

* Thyra::DefaultInverseModelEvaluator:  I added support for s.p.d. linear
operators Q_o and Q_p that are used as the scalar products in the matching
and regularization terms.  This allowed me implement the exact same objective
that is used in the GLpApp::AdvDiffReactOptModel class and allowed me to
maintain a bunch of different matching vectors needed for the multi-period
problem.  I also added a parameter list parameter to add the state observation
matching vector as an InArgs p(l) parameter.  This allowed me to have
a single Thyra::DefaultInverseModelEvaluator object for all of the periods
and then pass in the full state needed to for the evaluation.

* I added overrides for describe(...) and description() to
Thyra::DefaultProductVectorSpace to print out what the structure of the
product space is.  This is very useful in debugging the multi-period code.

* I have added support for a null domain space for the Epetra/Thyra wrapper
functions Thyra::create_MultiVector(...).  I made this change to make it
easier to create a Thyra::MultiVector wrapper for an Epetra_MultiVector
object.  I used this in the creation of trans(B_bar)*R*B_bar in the inverse
objective function for the GL problem.  I really need to simplify the use of
these wrappers by adding a function to create the locally replicated domain
space for multi-vectors and I need to change all of the interfaces to take
only the base Thyra interfaces and then do dynamic casting.  This will
greatly simplify use..

* For the example GLpApp::AdvDiffReactOptModel class, I added functions to
expose the data pool object and B_bar.  This allowed me to get at R, H, and
B_bar objects in order to build the external objective function with
Thyra::DefaultInverseModelEvaluator for use in the  so that when you rest the output context then
you wipe out all of the output streams and they get created again, with the
right extensions.  This allows my
MultiPeriodNLPThyraEpetraAdvDiffReactOptMain.cpp example program to create the
three sets of fwd-init, fwd, and inv files correctly.

* Bug 3305: I commented out the call to MPI_Op_free(...) that cases some
implementations of MPI to complain.  I will do this fix in the release
branch shortly.

* Teuchos::AbstractFactoryStd:  I create some nonmember constructors called
abstractFactoryStd(...) that return RCP objects.  The resulted in the
change of the name of an existing function that was used by
AbstractLinAlgPack::BasisSystemComposite (hence the change to this file).

* Thyra::ModelEvaluatorBase:  I added the same functions to the Derivative
class that I did for the EpetraExt classes suggested by Andy Salinger.
I have updated enough code to use this to insure that it gets tested.

* I added a Thyra::relErr(...) function for vectors to Thyra_TestingTools.hpp.
I used this to implement the error checks in the
MultiPeriodNLPThyraEpetraAdvDiffReactOptMain.cpp program.

* I cleaned up documentation and code formatting for a lot of other code.